Major refactor
Major refactor of the database system to remove unnecessary classes (at this point in time) and actually get system functional. Presently does not work, but does not explode either.
Please do not delete the main function in DerbyDatabase, this func. is used to build the tables. Without it, we cannot access our tables.