Users can now look at fractals on seperate page
Right now it only shows the name and it doesn't look pretty, but works
on a fundamental level. Users can add fractals to a list, and their
names are all displayed on the main page list. I'm not sure if the test
fractals will be in this commit, but there are a lot of test fractals
that need to be removed from the database, but for now they act as some
basic test data.